rename and reorganise bundles
