+++ packages/ag-charts-enterprise/src/module-bundles/cartesian-axes.ts
@@ -1,0 +1,9 @@
+import { AllCartesianAxesModule as AllCommunityCartesianAxesModule } from 'ag-charts-community';
+import type { ModuleDefinition } from 'ag-charts-core';
+
+import { OrdinalTimeAxisModule } from '../axes/ordinal/ordinalTimeAxisModule';
+
+export const AllCartesianAxesModule: ModuleDefinition[] = [
+    AllCommunityCartesianAxesModule,
+    OrdinalTimeAxisModule,
+].flat();
