+++ packages/ag-charts-enterprise/src/module-bundles/gauge.ts
@@ -1,0 +1,6 @@
+import type { ModuleDefinition } from 'ag-charts-core';
+
+import { LinearGaugeModule } from '../series/linear-gauge/linearGaugeModule';
+import { RadialGaugeModule } from '../series/radial-gauge/radialGaugeModule';
+
+export const AllGaugeModule: ModuleDefinition[] = [LinearGaugeModule, RadialGaugeModule];
